+++ meta_package_manager/__init__.py
@@ -24,7 +24,7 @@ import logging
 
 from .bitbar import fix_environment
 
-__version__ = '2.10.0'
+__version__ = '3.0.0'
 """ Examples of valid version strings according :pep:`440#version-scheme`:
 
 .. code-block:: python