+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 
 .. note:: This version is not yet released and is under active development.
 
+* [NPM]: 'install package@version' instead of 'update package’
+* [NPM]: skip update notifier
 * Fallback on ``apt version apt`` when looking for version. Closes #57 and #52.