+++ CHANGES.rst
@@ -17,6 +17,7 @@ Changelog
 * [NPM]: skip update notifier
 * [brew,cask]: Allow independent search for each manager.
 * [brew,cask]: Bump minimal requirement of to ``2.2.9``.
+* [mpm]: Do not force sync when calling outdated. Closes #36.
 * Fallback on ``apt version apt`` when looking for version. Closes #57 and #52.
 * Raise requirement to ``boltons >= 20.0``.