+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-`2.10.0 (unreleased) <https://github.com/kdeldycke/meta-package-manager/compare/v2.9.0...develop>`_
+`3.0.0 (unreleased) <https://github.com/kdeldycke/meta-package-manager/compare/v2.9.0...develop>`_
 --------------------------------------------------------------------------------------------------
 
 .. note:: This version is not yet released and is under active development.